requirements.txt
requests
python-dotenv
pydantic>=2.0.0
httpx>=0.24.0
